Make Elevation floating point.

+++ code/primitives.h
@@ -36,7 +36,7 @@ typedef uint16_t uint16;
 typedef uint32_t uint32;
 typedef uint64_t uint64;
 
-typedef int16 Elevation;
+typedef float Elevation;
 typedef int32 Coord;
 
 // Encoded x, y offsets inside tile
